feat: project convert
convert command for cac
